+++ include/cuml4c/kmeans.h
@@ -9,8 +9,8 @@
 
 EXTERN_C int KmeansFit(
     const float *x,
-    int num_row,
-    int num_col,
+    size_t num_row,
+    size_t num_col,
     const float *sample_weight,
     int k,
     int max_iters,
